backend/requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
numpy==1.26.0
requests==2.30.0
flask-cors==4.0.0